feat: support expo config plugin (#25)
To make it easier for people using this library with Expo, I added support for the Expo config plugin to enable automatic generation of native configurations. This feature is currently in the experimental phase, so any feedback is greatly appreciated!

Simply add the config plugin to your `app.json` file and specify the list of supported languages:

```json
{
  "expo": {
    // ...
    "plugins": [
      [
        "react-native-localization-settings", 
        { 
          "languages": ["en", "pl"] 
        }
      ]
    ]
  }
}
```

and run `npx expo prebuild`.